Number Generation
A random number is now assigned to each user in a comment thread. OP has
the number 0, and everyone else in the thread is given a unique number [
1, 100 ]. Going to add the functionality to prevent duplicate user
numbers on the next commit here in a few minutes. Just realized right
now that I forgot to do that last week when I wrote this.