First attempt at location in posts
Adds location variable, which defaults to "GLOBAL". Watches the user's location (if they have location off or deny request for location it just stays "GLOBAL") and will pass GPS coordinates to google's geocoding api to get the neighborhood. If it's anything out of oakland, location remains "GLOBAL", otherwise it's set to the correcty subset of oakland.