images
I went back to my apt to push this cause I wasn't sure when I was
leaving the lab and got surprise merge conflicts lol. That's what I get
for doing this last week and never pushing it. I see Christian fixed my
one problem. I changed it to my solution for clarity for an external
viewer (storing the list as post.sub_comments makes it seem like
sub_comments is contained within post and not generated).  Basically the
rest of my work for this commit was the effort of vectorizing and
recoloring the icons. Livetrace-ing the icons in AI looked like doodoo
so I did it by hand. 10 images in 10 colors and 1 OP, 101 images total.
Added the line of text to the js file and the JSON creation to print
these out.